requirements.txt
# RSS解析
feedparser>=6.0.0

# HTTP请求
requests>=2.28.0

# 模板引擎
Jinja2>=3.1.0

# 日期处理
python-dateutil>=2.8.0

# YAML解析
PyYAML>=6.0

# HTTP客户端（备用）
httpx>=0.24.0

# 图片处理（可选）
Pillow>=9.0.0
